+++ example/lib/model/main_item.dart
@@ -4,3 +4,10 @@ class MainItem {
   String description;
   String route;
 }
+
+class MainRouteItem {
+  MainRouteItem(this.title, this.description, {this.route});
+  String title;
+  String description;
+  MainRouteItem route;
+}